path: fixed MANPATH setup
The problem was that the man pages for the Perl modules I had
installed via Perlbrew weren't showing up. The problem was that
MANPATH wasn't set correctly. Once fixed, the next problem was that
completion didn't show the newly accessible man pages. Here, the
problem was that man queries the index, which is created by mandb,
which runs without the MANPATH setting, and thus ignores the man pages
for the Perl modules I had installed. I added a defun to run mandb for
me, in order to remind me of the problem.